feat(payments): marcar cita como Atendida al registrar pago
Al crear un pago se marca la cita como Atendida y se considera Atendida como estado finalizado para el calculo de comisiones. Agrega migracion de estado Atendida y ajustes menores (tipos, tests, imports).

+++ src/files/infrastructure/files.service.ts
@@ -48,7 +48,9 @@ export class FilesService {
     const chunks: Buffer[] = [];
 
     for await (const chunk of file.body) {
-      chunks.push(Buffer.isBuffer(chunk) ? chunk : Buffer.from(chunk));
+      chunks.push(
+        Buffer.isBuffer(chunk) ? chunk : Buffer.from(chunk as Uint8Array),
+      );
     }
 
     return {
